commander-core/requirements.txt
# Core
python-dotenv==1.0.1
APScheduler==3.10.4
pydantic==2.8.2
bandit==1.7.9

# AI & Routing
litellm==1.41.24
# Replace with actual image/music generation libraries if using official SDKs
# openai==1.35.10 (if needed natively outside litellm)

# Memory & Vector Search (PostgreSQL + numpy cosine similarity)
numpy
Pillow
duckduckgo-search
langgraph>=0.2.20
langchain-core>=0.2.30
langchain-openai>=0.1.20
# Parsing
tree-sitter==0.22.3
# Usually needs language specific bindings like tree-sitter-python, tree-sitter-typescript
tree-sitter-python==0.21.0
tree-sitter-typescript==0.21.2

# API & Web (if exposing internal API)
fastapi==0.111.0
uvicorn==0.30.1
requests==2.32.3
beautifulsoup4==4.12.3
PyYAML
sqlalchemy==2.0.35
psycopg2-binary==2.9.9

# Storage (Cloudflare R2 is S3 compatible)
boto3==1.34.144

# Communication (Telegram)
aiogram==3.10.0

# Finance (CFO module)
stripe==10.4.0

# Vector DB (now using pgvector via PostgreSQL, no separate service needed)

# Advanced Web Scraping (Clawbot)
crawl4ai==0.3.3
playwright>=1.44.0
markdown-it-py>=3.0.0
